Refactor: move ClickContext and ClickProperties into the root ui package
