+++ src/main/java/co/crystaldev/alpinecore/framework/ui/ClickContext.java
@@ -1,4 +1,4 @@
-package co.crystaldev.alpinecore.framework.ui.element;
+package co.crystaldev.alpinecore.framework.ui;
 
 import co.crystaldev.alpinecore.framework.ui.event.ActionResult;
 import org.bukkit.Material;
