+++ src/main/java/co/crystaldev/alpinecore/framework/ui/ClickFunction.java
@@ -1,6 +1,5 @@
 package co.crystaldev.alpinecore.framework.ui;
 
-import co.crystaldev.alpinecore.framework.ui.element.ClickContext;
 import org.jetbrains.annotations.NotNull;
 
 /**
